pyproject.toml
[project]
name = "agent-repl"
version = "0.1.0"
description = "CLI tool giving AI agents direct access to live Jupyter notebook kernels"
requires-python = ">=3.10"
dependencies = [
    "jupyter-client>=8",
    "jupyter-server>=2",
    "nbformat>=5",
    "requests>=2",
    "websocket-client>=1.8",
]

[project.scripts]
agent-repl = "agent_repl.cli:main_entry"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_repl"]

[dependency-groups]
dev = [
    "pytest",
]
